% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trelliscope_display.R
\name{build_trelliscope_display}
\alias{build_trelliscope_display}
\title{Generate a trelliscope display for all Proteoform to MS1 Matches}
\usage{
build_trelliscope_display(
  PeakData,
  Ms1Match,
  Path = file.path(.getDownloadsFolder(), "Ms1Match"),
  Window = 2,
  ...
)
}
\arguments{
\item{PeakData}{A pspecterlib peak_data object or data.table with "M/Z" and "Intensity". Required.}

\item{Ms1Match}{A ProteoMatch_MatchedPeaksclass object from match_full_seq_ms1. Required.}

\item{Path}{The base directory of the trelliscope application. Default is Downloads/Ms1Match.}

\item{Window}{The -/+ m/z value on either side of the matched spectra. Default is 2 m/z.}
}
\value{
An html trelliscope display
}
\description{
\href{https://github.com/hafen/trelliscopejs}{Trelliscope} allows
users to filter and sort plots based on cognostic variables. For this dataset,
the plots are generated with the plot_Ms1Match function, and the cognostics
are: Protein, Absolute Relative Error, Correlation, Charge, Proteoform, and ID.
}
