% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_molform.R
\name{calculate_molform}
\alias{calculate_molform}
\title{Generate a Molecular Formula Table from proteoforms}
\usage{
calculate_molform(
  Proteoform,
  Protein = NULL,
  Charge = 1,
  ProtonMass = 1.00727647
)
}
\arguments{
\item{Proteoform}{A vector of strings written in proteoform annotation. i.e. "M.AA\code{[}Acetyl\verb{]}AA\code{[}3.2\verb{]}.V"}

\item{Protein}{A vector of identifiers for each protein. Default is NULL.}

\item{Charge}{The range of charges to test. Default is 1.}

\item{ProtonMass}{The AMU mass of a proton. Default is 1.00727647.}
}
\value{
A list containing the molecular formula, mass shift, monoisotopic mass,
and most abundant isotope
}
\description{
Returns the "ProteoMatch_molform" object with each proteoform's
molecular formula and adjusted mass.
}
\details{
The data.table outputted by this function returns 8 columns.
\tabular{ll}{
Molecular Formula \tab The molecular formula derived from the peptide/protein sequence and its modifications \cr
\tab \cr
Mass Shift \tab The total mass change indicated in the proteoform \cr
\tab \cr
Monoisotopic Mass \tab The proteoform's monoisotopic peak based on molecular formula and mass shift \cr
\tab \cr
Most Abundant Isotope \tab The proteoform's most abundant isotope based on molecular formula and mass shift \cr
\tab \cr
Protein \tab The provided protein identifier \cr
\tab \cr
Proteoform \tab The provided proteoform \cr
\tab \cr
}
}
