% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_full_seq_ms1.R
\name{match_full_seq_ms1}
\alias{match_full_seq_ms1}
\title{Matches peaks to isotoping profile, given the molecular formula and any mass changes}
\usage{
match_full_seq_ms1(
  PeakData,
  MolecularFormula,
  MassShift = NULL,
  Charge = 1,
  IsotopicPercentage = 1,
  PPMThreshold = 10,
  MaxIsotopes = 20
)
}
\arguments{
\item{PeakData}{A pspecterlib peak_data object or data.table with "M/Z" and "Intensity". Required.}

\item{MolecularFormula}{The molecular formula written as string with element, numeric.
It should include the modifications' molecular formula calculated in calculate_molform. Required.}

\item{MassShift}{Any changes to the mass written as a numeric. Default is NULL}

\item{Charge}{The charge of the mass. Numeric. Default is 1.}

\item{IsotopicPercentage}{The minimum isotopic percentage (calculated intensity) permitted
to be matched. Default is 0.1, which means 0.1\%.}

\item{PPMThreshold}{The maximum m/z error permitted. Default is 10 ppm.}

\item{MaxIsotopes}{The maximum number of isotopes to consider. Default is 30.}
}
\description{
Matches peaks to isotoping profile, given the molecular formula and any mass changes
}
