% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_proteoform_to_ms1.R
\name{match_proteoform_to_ms1}
\alias{match_proteoform_to_ms1}
\title{Matches experimental peaks to calculated isotope profile}
\usage{
match_proteoform_to_ms1(
  PeakData,
  MolecularFormula,
  IsotopicPercentage = 1,
  PPMThreshold = 10,
  MaxIsotopes = 20
)
}
\arguments{
\item{PeakData}{A pspecterlib peak_data object made with "make_peak_data" or
extracted from an mzML file with "get_peak_data." Use centroided data. Required.}

\item{MolecularFormula}{A "ProteoMatch_MolForm" object with Molecular Formulas,
Mass Shifts, and Charges.}

\item{IsotopicPercentage}{The minimum isotopic percentage (calculated intensity) permitted
to be matched. Default is 0.1, which means 0.1\%.}

\item{PPMThreshold}{The maximum m/z error permitted. Default is 10 ppm.}

\item{MaxIsotopes}{The maximum number of isotopes to consider. Default is 30.}
}
\value{
A ProteoMatch_MatchedPeaks object, which is a data.table containing the
Protein, M/Z, Intensity, Isotope, M/Z Tolerance, Experimental M/Z,
Experimental Intensity, PPM Error, Absolute Relative Error, Correlation,
Charge, and Proteoform.
}
\description{
Returns the "ProteoMatch_MatchedPeaks" object with peaks that
have matched to the isotoping profile. This object can be passed to downstream
plotting functions.
}
\details{
The data.table outputted by this function contains 12 columns
\tabular{ll}{
Protein \tab The provided protein identifier \cr
\tab \cr
M/Z \tab The calculated M/Z of a particular isotope \cr
\tab \cr
Intensity \tab The calculated relative intensity of a particular isotope \cr
\tab \cr
Isotope \tab The isotope number "n" in M+n \cr
\tab \cr
M/Z Tolerance \tab Given the ppm, the M/Z tolerance is the window in which peaks are searched and matched \cr
\tab \cr
M/Z Experimental \tab The M/Z of the matched from the experimental peaks \cr
\tab \cr
Intensity Experimental \tab The intensity of the matched peak from the experimental peaks \cr
\tab \cr
PPM Error \tab How far off the calculated peak is from the experimental peak. Only peaks within the ppm threshold are kept \cr
\tab \cr
Absolute Relative Error \tab The sum of the absolute relative difference of calculated and experimental peak intensities. \cr
\tab \cr
Correlation \tab The cosine correlation of calculated and experimental peak intensities \cr
\tab \cr
Charge \tab The provided charges \cr
\tab \cr
Proteoform \tab The provided proteoform \cr
\tab \cr
}
}
