% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ms1_match.R
\name{plot_ms1_match}
\alias{plot_ms1_match}
\title{Matches peaks to isotoping profile, given the molecular formula and any mass changes}
\usage{
plot_ms1_match(PeakData, ms1_match)
}
\arguments{
\item{PeakData}{A pspecterlib peak_data object or data.table with "M/Z" and "Intensity". Required.}

\item{ms1_match}{A ms1_match class object from match_full_seq_ms1. Required.}
}
\description{
Matches peaks to isotoping profile, given the molecular formula and any mass changes
}
