% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_proteomatch.R
\name{run_proteomatch}
\alias{run_proteomatch}
\title{Run the ProteoMatch pipeline with a Proteoform, mzML, Settings, and Defaults File}
\usage{
run_proteomatch(ProteoformFile, mzMLFile, SettingsFile)
}
\arguments{
\item{ProteoformFile}{Path to a csv with a column of "Proteoforms" and their "Protein"
identifier. Each Proteoform-Protein pairing should be unique. Proteoforms
should be written with proper annotations (i.e. "M.AA\code{[}Acetyl\verb{]}AA\code{[}3.2\verb{]}.V"). Required.}

\item{mzMLFile}{Path to an mzML file containing one spectra. The file must be centroided.
Use the msconverteR \url{https://github.com/wilsontom/msconverteR} package and
set}
}
\description{
A wrapper for quickly running all the steps of the ProteoMatch pipeline
}
